자료형 통일

+++ src/main/java/project/maybedo/domain/Message.java
@@ -2,7 +2,7 @@ package project.maybedo.domain;
 
 import lombok.Getter;
 import lombok.Setter;
-import project.maybedo.group.Group;
+import project.maybedo.group.domain.Group;
 import project.maybedo.member.Member;
 
 import javax.persistence.*;
@@ -14,7 +14,7 @@ public class Message {
     @Id
     @GeneratedValue
     @Column(name = "message_id")
-    private Long id;
+    private int id;
     private String content;
     private LocalDateTime time;
